Fixeed an issue with moving not properly working.
git-svn-id: https://109.226.161.129/svn/code/java/utiLITI/trunk@3551 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/utiLITI/Components/GridEditPanel.java
@@ -47,6 +47,6 @@ public class GridEditPanel extends JPanel {
   }
 
   public int getGridSize() {
-    return (int) (long) this.textField.getValue();
+    return (int) this.textField.getValue();
   }
 }
